// lib/rest-api/types.d.ts
export interface AxiosRequestArgs {
    url: string;
    options: RawAxiosRequestConfig;
}
/**
 * Represents the response from a REST API request.
 * @template T - The type of the data returned in the response.
 * @property {() => Promise<T>} data - A function that returns a Promise resolving to the data from the API response.
 * @property {number} status - The HTTP status code of the response.
 */
export type RestApiResponse<T> = {
    data: () => Promise<T>;
    status: number;
};
export declare class ConfigurationRestAPI {
    /**
     * override base path
     * @type {string}
     * @memberof ConfigurationRestAPI
     */
    basePath?: string;
    /**
     * set a timeout (in milliseconds) for the request
     * @default 1000
     * @type {number}
     * @memberof ConfigurationRestAPI
     */
    timeout?: number;
    /**
     * number of retry attempts for failed requests
     * @default 1
     * @type {number}
     * @memberof ConfigurationRestAPI
     */
    retries?: number;
    /**
     * delay between retry attempts in milliseconds
     * @default 1000
     * @type {number}
     * @memberof ConfigurationRestAPI
     */
    backoff?: number;
    /**
     * base options for axios calls
     * @type {Record<string, unknown>}
     * @memberof ConfigurationRestAPI
     * @internal
     */
    baseOptions?: Record<string, unknown>;
    constructor(param: ConfigurationRestAPI);
}